feat: Implement TableRequirement checks (#689)
* Impelment TableRequirement check

* Address comments

+++ crates/iceberg/src/transaction.rs
@@ -154,7 +154,7 @@ impl<'a> ReplaceSortOrderAction<'a> {
 
         let requirements = vec![
             TableRequirement::CurrentSchemaIdMatch {
-                current_schema_id: self.tx.table.metadata().current_schema().schema_id() as i64,
+                current_schema_id: self.tx.table.metadata().current_schema().schema_id(),
             },
             TableRequirement::DefaultSortOrderIdMatch {
                 default_sort_order_id: self.tx.table.metadata().default_sort_order().order_id,
